API/requirements.txt
flask==3.0.0
requests
flask-cors
python-dotenv==1.0.0
